change Login structure
