+++ src/Components/Shared/CarouselAutoPlay/CarouselAutoPlayView.tsx
@@ -14,7 +14,6 @@ export default () => {
     <a href="https://mhfd.org/" target="_blank">
       <div className="logo-white"
         style={{ backgroundImage: 'url(/Icons/logo-white.svg)' }}>
-        {/*<p>Protecting People, Property, and our Environment.</p>*/}
       </div>
     </a>
     <div className="contact01">
@@ -37,15 +36,6 @@ export default () => {
           <Link href="#login-form" className="btn-purple" title="Enter"/>
         </Anchor>
       </div>
-    </div>
-    <div className="logo-partner">
-      {/* <div>
-         <p>Designed by</p>
-        <a href="https://vizonomy.com/main" target="_blank"><img className="vizonomy" src="/Icons/logo_vizonomy.png" alt="" /></a> 
-      </div>
-      <div>
-        <a href="https://www.respec.com" target="_blank"><img className="respec" src="/Icons/logo-respec.png" alt="" /></a>
-      </div> */}
     </div>
     <Carousel autoplay autoplaySpeed={6000}>
       {images.map((image: string, index: number) => {
