+++ src/routes/login/components/VersionNumber.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const VersionNumber = () => {
+  return (
+    <span style={{ position: 'absolute', bottom: '5px', marginLeft: '10px' }}>1.4.12</span>
+  )
+};
+
+export default VersionNumber;
