+++ src/routes/login/index.ts
@@ -1,0 +1,3 @@
+import LoginProvider from './components/LoginProvider';
+
+export default LoginProvider;
